requirements.txt
django==3.1.7
django-ckeditor==6.0.0
Pillow==8.1.2